requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0.5
django-filebrowser==3.13.1
django-froala-editor==3.1.0
django-grappelli==2.14.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
psycopg2==2.8.5
pylint==2.5.2
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
